Fix dangerous casts
On Windows 64-bit, the size of long is 32-bit, thus any pointer to long cast is
not valid. In any case, one should use intptr_t and ptrdiff_t when casting
to integers. This commit attempts to fix all instances reported by GCC.
When relevant, I replaced code by the macros PTR_ADD, ALIGN_UP from system.h

+++ apps/plugins/doom/d_deh.c
@@ -2082,7 +2082,7 @@ void deh_procSounds(DEHFILE *fpin, int fpout, char *line)
                S_sfx[indexnum].priority = (int)value;
             else
                if (!strcasecmp(key,deh_sfxinfo[3]))  // Zero 1
-                  S_sfx[indexnum].link = (sfxinfo_t *)((long) value);
+                  S_sfx[indexnum].link = (sfxinfo_t *)((intptr_t) value);
                else
                   if (!strcasecmp(key,deh_sfxinfo[4]))  // Zero 2
                      S_sfx[indexnum].pitch = (int)value;
@@ -2091,7 +2091,7 @@ void deh_procSounds(DEHFILE *fpin, int fpout, char *line)
                         S_sfx[indexnum].volume = (int)value;
                      else
                         if (!strcasecmp(key,deh_sfxinfo[6]))  // Zero 4
-                           S_sfx[indexnum].data = (void *) ((long) value); // killough 5/3/98: changed cast
+                           S_sfx[indexnum].data = (void *) ((intptr_t) value); // killough 5/3/98: changed cast
                         else
                            if (!strcasecmp(key,deh_sfxinfo[7]))  // Neg. One 1
                               S_sfx[indexnum].usefulness = (int)value;
